Refactor users table
