+++ packages/data-access-plugin/src/ui/components/UsersTable/index.tsx
@@ -1,167 +1,17 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import classNames from 'classnames';
 import Table from '@material-ui/core/Table';
-import Tooltip from '@material-ui/core/Tooltip';
 import Checkbox from '@material-ui/core/Checkbox';
 import { AnyObject, get, uniq } from '@terascope/utils';
 import TableRow from '@material-ui/core/TableRow';
 import TableBody from '@material-ui/core/TableBody';
-import TableHead from '@material-ui/core/TableHead';
-import TableSortLabel from '@material-ui/core/TableSortLabel';
 import TablePagination from '@material-ui/core/TablePagination';
-import { lighten } from '@material-ui/core/styles/colorManipulator';
 import TableCell, { SortDirection } from '@material-ui/core/TableCell';
 import { Theme, createStyles, withStyles } from '@material-ui/core/styles';
-import { ResolvedUser, QueryState } from '../interfaces';
-import Toolbar from '@material-ui/core/Toolbar';
-import Typography from '@material-ui/core/Typography';
-import IconButton from '@material-ui/core/IconButton';
-import DeleteIcon from '@material-ui/icons/Delete';
-import FilterListIcon from '@material-ui/icons/FilterList';
-
-const rows = [
-  { id: 'firstname', label: 'First Name' },
-  { id: 'lastname', label: 'Last Name' },
-  { id: 'username', label: 'Username' },
-  { id: 'role', label: 'Role' },
-  { id: 'created', label: 'Created' },
-];
-
-type EnhancedTableHeadProps = {
-    numSelected: number;
-    onRequestSort: (event: any, property: string) => void;
-    onSelectAllClick: (event: any) => void;
-    order: SortDirection;
-    orderBy: string;
-    rowCount: number;
-};
-
-class EnhancedTableHead extends React.Component<EnhancedTableHeadProps> {
-    static propTypes = {
-        numSelected: PropTypes.number.isRequired,
-        onRequestSort: PropTypes.func.isRequired,
-        onSelectAllClick: PropTypes.func.isRequired,
-        order: PropTypes.string.isRequired,
-        orderBy: PropTypes.string.isRequired,
-        rowCount: PropTypes.number.isRequired,
-    };
-
-    createSortHandler = property => event => {
-        this.props.onRequestSort(event, property);
-    }
-
-    render() {
-        const { onSelectAllClick, order, orderBy, numSelected, rowCount } = this.props;
-
-        return (
-            <TableHead>
-                <TableRow>
-                    <TableCell padding="checkbox">
-                        <Checkbox
-                            indeterminate={numSelected > 0 && numSelected < rowCount}
-                            checked={numSelected === rowCount}
-                            onChange={onSelectAllClick}
-                        />
-                    </TableCell>
-                    {rows.map((row) => (<TableCell
-                            key={row.id}
-                            align="left"
-                            sortDirection={orderBy === row.id ? order : false}
-                        >
-                        <Tooltip
-                            title="Sort"
-                            placement="bottom-start"
-                            enterDelay={300}
-                        >
-                            <TableSortLabel
-                                active={orderBy === row.id}
-                                direction={order as 'asc'|'desc'}
-                                onClick={this.createSortHandler(row.id)}
-                            >
-                                {row.label}
-                            </TableSortLabel>
-                        </Tooltip>
-                    </TableCell>), this)}
-                </TableRow>
-            </TableHead>
-        );
-    }
-}
-
-const toolbarStyles = theme => ({
-    root: {
-        paddingRight: theme.spacing.unit,
-    },
-    highlight:
-    theme.palette.type === 'light'
-      ? {
-          color: theme.palette.secondary.main,
-          backgroundColor: lighten(theme.palette.secondary.light, 0.85),
-      }
-      : {
-          color: theme.palette.text.primary,
-          backgroundColor: theme.palette.secondary.dark,
-      },
-    spacer: {
-        flex: '1 1 100%',
-    },
-    actions: {
-        color: theme.palette.text.secondary,
-    },
-    title: {
-        flex: '0 0 auto',
-    },
-});
-
-class EnhancedTableToolbar extends React.Component<{ numSelected: number, classes: any }> {
-    static propTypes = {
-        classes: PropTypes.object.isRequired,
-        numSelected: PropTypes.number.isRequired,
-    };
-
-    render() {
-        const { numSelected, classes } = this.props;
-
-        return (
-            <Toolbar
-                className={classNames(classes.root, {
-                    [classes.highlight]: numSelected > 0,
-                })}
-            >
-                <div className={classes.title}>
-                    {numSelected > 0 ? (
-                    <Typography color="inherit" variant="subtitle1">
-                        {numSelected} selected
-                    </Typography>
-                    ) : (
-                    <Typography variant="h6" id="tableTitle">
-                        Nutrition
-                    </Typography>
-                    )}
-                </div>
-                <div className={classes.spacer} />
-                <div className={classes.actions}>
-                    {numSelected > 0 ? (
-                    <Tooltip title="Delete">
-                        <IconButton aria-label="Delete">
-                        <DeleteIcon />
-                        </IconButton>
-                    </Tooltip>
-                    ) : (
-                    <Tooltip title="Filter list">
-                        <IconButton aria-label="Filter list">
-                        <FilterListIcon />
-                        </IconButton>
-                    </Tooltip>
-                    )}
-                </div>
-            </Toolbar>
-        );
-    }
-}
-
-const EnhancedTableToolbarWithStyles = withStyles(toolbarStyles)(EnhancedTableToolbar);
+import { ResolvedUser, QueryState } from '../../interfaces';
+import UsersTableToolbar from './toolbar';
+import UsersTableHeader from './header';
+import { rows } from './shared';
 
 const styles = (theme: Theme) => createStyles({
     root: {
@@ -306,9 +156,9 @@ class Users extends React.Component<UsersTableProps, TableState> {
 
         return (
             <div className={classes.tableWrapper}>
-                <EnhancedTableToolbarWithStyles numSelected={selected.length} />
+                <UsersTableToolbar numSelected={selected.length} />
                 <Table className={classes.table}>
-                    <EnhancedTableHead
+                    <UsersTableHeader
                         numSelected={selected.length}
                         order={order}
                         orderBy={orderBy}
