+++ packages/data-access-plugin/src/ui/components/UsersTable/shared.ts
@@ -1,0 +1,7 @@
+export const rows = [
+  { id: 'firstname', label: 'First Name' },
+  { id: 'lastname', label: 'Last Name' },
+  { id: 'username', label: 'Username' },
+  { id: 'role', label: 'Role' },
+  { id: 'created', label: 'Created' },
+];
