+++ packages/data-access-plugin/src/ui/components/UsersTable/toolbar.tsx
@@ -1,0 +1,85 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import classNames from 'classnames';
+import Tooltip from '@material-ui/core/Tooltip';
+import Toolbar from '@material-ui/core/Toolbar';
+import DeleteIcon from '@material-ui/icons/Delete';
+import Typography from '@material-ui/core/Typography';
+import IconButton from '@material-ui/core/IconButton';
+import FilterListIcon from '@material-ui/icons/FilterList';
+import { lighten } from '@material-ui/core/styles/colorManipulator';
+import { Theme, createStyles, withStyles } from '@material-ui/core/styles';
+
+const toolbarStyles = (theme: Theme) => createStyles({
+    root: {
+        paddingRight: theme.spacing.unit,
+    },
+    highlight:
+    theme.palette.type === 'light'
+      ? {
+          color: theme.palette.secondary.main,
+          backgroundColor: lighten(theme.palette.secondary.light, 0.85),
+      }
+      : {
+          color: theme.palette.text.primary,
+          backgroundColor: theme.palette.secondary.dark,
+      },
+    spacer: {
+        flex: '1 1 100%',
+    },
+    actions: {
+        color: theme.palette.text.secondary,
+    },
+    title: {
+        flex: '0 0 auto',
+    },
+});
+
+class UsersTableToolbar extends React.Component<{ numSelected: number, classes: any }> {
+    static propTypes = {
+        classes: PropTypes.object.isRequired,
+        numSelected: PropTypes.number.isRequired,
+    };
+
+    render() {
+        const { numSelected, classes } = this.props;
+
+        return (
+            <Toolbar
+                className={classNames(classes.root, {
+                    [classes.highlight]: numSelected > 0,
+                })}
+            >
+                <div className={classes.title}>
+                    {numSelected > 0 ? (
+                    <Typography color="inherit" variant="subtitle1">
+                        {numSelected} selected
+                    </Typography>
+                    ) : (
+                    <Typography variant="h6" id="tableTitle">
+                        Users
+                    </Typography>
+                    )}
+                </div>
+                <div className={classes.spacer} />
+                <div className={classes.actions}>
+                    {numSelected > 0 ? (
+                    <Tooltip title="Delete">
+                        <IconButton aria-label="Delete">
+                            <DeleteIcon />
+                        </IconButton>
+                    </Tooltip>
+                    ) : (
+                    <Tooltip title="Filter list">
+                        <IconButton aria-label="Filter list">
+                            <FilterListIcon />
+                        </IconButton>
+                    </Tooltip>
+                    )}
+                </div>
+            </Toolbar>
+        );
+    }
+}
+
+export default withStyles(toolbarStyles)(UsersTableToolbar);
